// src/api/models/article.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * Aris-blog
 * OpenAPI spec version: 1.0
 */
import type { ArticleTags } from './articleTags';
import type { User } from './user';

export interface Article {
  /**
   * Article ID
   * @minimum 0
   */
  articleID: number;
  /** Number of comments */
  comments: number;
  /** Creation timestamp */
  createdAt: string;
  /**
   * Number of likes
   * @minimum 0
   */
  likes: number;
  /** Publication timestamp */
  publishedAt: string;
  /** Article slug */
  slug: string;
  /** Article status */
  status: string;
  /** List of tags */
  tags: ArticleTags;
  /** Article title */
  title: string;
  /** Update timestamp */
  updatedAt: string;
  /** Author information */
  user: User;
  /**
   * Number of views
   * @minimum 0
   */
  views: number;
}